+++ tests/test_models/test_linear_model.py
@@ -52,6 +52,18 @@ def linear_segments_ts_common():
     return linear_segments_by_parameters(alpha_values, intercept_values)
 
 
+@pytest.mark.parametrize("model", (LinearPerSegmentModel(), ElasticPerSegmentModel()))
+def test_not_fitted(model, linear_segments_ts_unique):
+    """Check exception when trying to forecast with unfitted model."""
+    train, test = linear_segments_ts_unique
+    lags = LagTransform(in_column="target", lags=[3, 4, 5])
+    train.fit_transform([lags])
+
+    to_forecast = train.make_future(3)
+    with pytest.raises(ValueError, match="model is not fitted"):
+        model.forecast(to_forecast)
+
+
 @pytest.mark.parametrize(
     "model_class, model_class_repr",
     ((LinearPerSegmentModel, "LinearPerSegmentModel"), (LinearMultiSegmentModel, "LinearMultiSegmentModel")),